requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.5
Django==3.2.7
django-bootstrap-forms==0.1
django-crispy-forms==1.12.0
django-filter==2.4.0
flake8==3.9.2
gunicorn==20.1.0
mccabe==0.6.1
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
